requirements.txt
apache-airflow
httpx
pandas